requirements.txt
pyyaml
unittest-xml-reporting
